feat(console): 迁 metric-chart 链到 console，完成 console 进程抽出
把 metric-chart 的 handler/service/dao + metric domain 整体迁入 @kagami/console（domain 自包含纯类型，server 侧留只含 MetricTags 的精简版，结构等价互通）。server 保留 metricService/metricDao 供 agent 打点；删除 metricChart 构造、handler 注册与 ServerRuntime 返回字段。web-server.mjs 把 /metric-chart 前缀也分流到 console。

修并发竞态：根 test 加 --workspace-concurrency=1，避免 console/server 两个 pretest 同时 prisma generate 撞 server-core 生成目录（EEXIST）。

本地冒烟：console /metric-chart/list 200。全门绿（console 9 测试文件/27 用例，server 612）。至此服务前端的 5 条查询链全部迁出 console，apps/server 不改名、agent 核心未动。
