+++ apps/console/src/app/console-runtime.ts
@@ -7,6 +7,7 @@ import { PrismaLogDao } from "@kagami/server-core/logger/dao/impl/log.impl.dao";
 import { PrismaLlmChatCallDao } from "@kagami/server-core/dao/impl/llm-chat-call.impl.dao";
 import { PrismaNapcatEventDao } from "@kagami/server-core/dao/impl/napcat-event.impl.dao";
 import { PrismaNapcatQqMessageDao } from "@kagami/server-core/dao/impl/napcat-group-message.impl.dao";
+import { PrismaMetricDao } from "@kagami/server-core/dao/impl/prisma-metric.impl.dao";
 import { BizError } from "@kagami/server-core/common/errors/biz-error";
 import { toHttpErrorResponse } from "@kagami/server-core/common/errors/http-error";
 import { AppLogger } from "@kagami/server-core/logger/logger";
@@ -16,10 +17,13 @@ import { AppLogHandler } from "../ops/http/app-log.handler.js";
 import { LlmChatCallHandler } from "../ops/http/llm-chat-call.handler.js";
 import { NapcatEventHandler } from "../ops/http/napcat-event.handler.js";
 import { NapcatQqMessageHandler } from "../ops/http/napcat-group-message.handler.js";
+import { MetricChartHandler } from "../metric/http/metric-chart.handler.js";
 import { DefaultAppLogQueryService } from "../ops/application/app-log-query.impl.service.js";
 import { DefaultLlmChatCallQueryService } from "../ops/application/llm-chat-call-query.impl.service.js";
 import { DefaultNapcatEventQueryService } from "../ops/application/napcat-event-query.impl.service.js";
 import { DefaultNapcatQqMessageQueryService } from "../ops/application/napcat-group-message-query.impl.service.js";
+import { DefaultMetricChartService } from "../metric/application/metric-chart.impl.service.js";
+import { PrismaMetricChartDao } from "../metric/infra/impl/prisma-metric-chart.impl.dao.js";
 
 const TRACE_ID_HEADER_NAME = "X-Kagami-Trace-Id";
 const logger = new AppLogger({ source: "console-bootstrap" });
@@ -63,6 +67,13 @@ export async function buildConsoleRuntime(): Promise<ConsoleRuntime> {
     napcatQqMessageDao,
   });
 
+  const metricDao = new PrismaMetricDao({ database });
+  const metricChartDao = new PrismaMetricChartDao({ database });
+  const metricChartService = new DefaultMetricChartService({
+    metricDao,
+    metricChartDao,
+  });
+
   const app = createConsoleApp({
     handlers: [
       new HealthHandler(),
@@ -70,6 +81,7 @@ export async function buildConsoleRuntime(): Promise<ConsoleRuntime> {
       new LlmChatCallHandler({ llmChatCallQueryService }),
       new NapcatEventHandler({ napcatEventQueryService }),
       new NapcatQqMessageHandler({ napcatQqMessageQueryService }),
+      new MetricChartHandler({ metricChartService }),
     ],
   });
