+++ apps/console/src/metric/application/metric-chart.impl.service.ts
