+++ apps/console/src/metric/application/metric-chart.service.ts
