+++ apps/console/src/metric/http/metric-chart.handler.ts
