+++ package.json
@@ -13,7 +13,7 @@
     "db:migrate:reset": "bash ./scripts/prisma.sh migrate reset",
     "db:migrate:resolve": "bash ./scripts/prisma.sh migrate resolve",
     "typecheck": "pnpm -r typecheck",
-    "test": "pnpm -r --if-present test",
+    "test": "pnpm -r --workspace-concurrency=1 --if-present test",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "format": "prettier . --check",
